requirements.txt
openai==0.27.4
python-dotenv==1.0.0
streamlit==1.20.0
streamlit_chat==0.0.2.2
